move crates/doc-poc => crates/doc
This commit is largely mechanical, with minor fixes to update call sites
for the new crate semantics.

`flowctl-admin combine` is substantially simplified with the switch to doc::combine.
