+++ crates/doc/Cargo.toml
@@ -10,19 +10,32 @@ license.workspace = true
 
 [dependencies]
 json = { path = "../json" }
+tuple = { path = "../tuple" }
 
-itertools = { workspace = true }
-lazy_static = { workspace = true }
+base64 = { workspace = true }
+bumpalo = { workspace = true }
 fancy-regex = { workspace = true }
+futures = { workspace = true }
+fxhash = { workspace = true }
+itertools = { workspace = true }
+rkyv = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+tempfile = { workspace = true }
 thiserror = { workspace = true }
-tinyvec = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
 
 [dev-dependencies]
+allocator = { path = "../allocator" }
+
+criterion = { workspace = true }
+hexdump = { workspace = true }
 insta = { workspace = true }
 quickcheck = { workspace = true }
 quickcheck_macros = { workspace = true }
+rand = { workspace = true }
+rand_distr = { workspace = true }
+serde-transcode = { workspace = true }
+serde_json = { workspace = true }
 serde_yaml = { workspace = true }
