+++ crates/doc/src/lazy.rs
