+++ crates/doc/src/lib.rs
@@ -1,102 +1,175 @@
+/// Node is the generic form of a document node as understood by Flow.
+/// It's implemented by HeapNode, ArchivedNode, and serde_json::Value.
+#[derive(Debug)]
+pub enum Node<'a, N: AsNode> {
+    Array(&'a [N]),
+    Bool(bool),
+    Bytes(&'a [u8]),
+    Null,
+    Number(json::Number),
+    Object(&'a N::Fields),
+    String(&'a str),
+}
+
+/// AsNode is the trait by which a specific document representation is accessed through a generic Node.
+pub trait AsNode: Sized {
+    type Fields: Fields<Self> + ?Sized;
+
+    fn as_node<'a>(&'a self) -> Node<'a, Self>;
+}
+
+/// Fields is the generic form of a Document object representation.
+pub trait Fields<N: AsNode> {
+    type Field<'a>: Field<'a, N>
+    where
+        Self: 'a;
+
+    type Iter<'a>: Iterator<Item = Self::Field<'a>>
+    where
+        Self: 'a;
+
+    fn get<'a>(&'a self, property: &str) -> Option<Self::Field<'a>>;
+    fn len(&self) -> usize;
+    fn iter<'a>(&'a self) -> Self::Iter<'a>;
+}
+
+/// Field is the generic form of a Document object Field as understood by Flow.
+pub trait Field<'a, N: AsNode> {
+    fn property(&self) -> &'a str;
+    fn value(&self) -> &'a N;
+}
+
+// This crate has three implementations of AsNode: a mutable HeapNode,
+// an ArchivedNode serialized by the `rkyv` crate,
+// and an implementation upon serde_json::Value.
+mod archived;
+pub use archived::{ArchivedDoc, ArchivedField, ArchivedNode};
+pub mod heap;
+pub use heap::{HeapDoc, HeapField, HeapNode};
+mod value;
+
+// Dedup de-duplicates strings used in the construction of HeapNodes.
+// This also reduces the size of serialized ArchivedNodes, as the archival format
+// stores one copy of each deduplicated string.
+pub mod dedup;
+
+// HeapNode may be directly deserialized using serde.
+mod heap_de;
+
+// We provide serde::Serialize covering all Doc implementations.
+mod ser;
+
+// All implementations of AsNode may be compared with one another.
+mod compare;
+pub use compare::compare;
+
+// A JSON Pointer implementation that works with all AsNode implementations,
+// and allows creation of documents using serde_json::Value and HeapNode.
 pub mod ptr;
-mod varint;
 pub use ptr::Pointer;
 
-pub mod inference;
-pub mod reduce;
+// Walker is a medium-term integration joint between AsNode implementations
+// and our JSON-schema validator. We may seek to get rid of this and have
+// JSON-schema validation evaluate directly over AsNode.
+pub mod walker;
 
+// Optimized conversions from AsNode implementations into HeapNode.
+pub mod lazy;
+pub use lazy::LazyNode;
+
+// JSON-schema annotation extensions supported by Flow documents.
 mod annotation;
 pub use annotation::Annotation;
 
-// Specialize json templates for the estuary `Annotation` type.
-pub type Schema = json::schema::Schema<Annotation>;
-pub type SchemaIndexBuilder<'sm> = json::schema::index::IndexBuilder<'sm, Annotation>;
-pub type SchemaIndex<'sm> = json::schema::index::Index<'sm, Annotation>;
-pub type FullContext = json::validator::FullContext;
-pub type SpanContext = json::validator::SpanContext;
-pub type Validator<'sm> = json::validator::Validator<'sm, Annotation, SpanContext>;
+// Validation is a higher-order API for driving JSON-schema validation
+// over AsNode implementations.
+pub mod validation;
+pub use validation::{
+    FailedValidation, Schema, SchemaIndex, SchemaIndexBuilder, Valid, Validation, Validator,
+};
 
-mod diff;
-pub use diff::Diff;
+// Doc implementations may be reduced.
+pub mod reduce;
 
-/// Validation represents the outcome of a document validation.
-pub struct Validation<'sm, 'v> {
-    /// Schema which was validated.
-    pub schema: &'v url::Url,
-    /// Document which was validated.
-    pub document: serde_json::Value,
-    /// Validator which holds the validation outcome.
-    // Note use of Validator in a loop requires that we separate these lifetimes.
-    pub validator: &'v mut Validator<'sm>,
-    /// Walked document span.
-    pub span: json::Span,
-}
+// Documents may be combined.
+pub mod combine;
+pub use combine::Combiner;
 
-// Validation is a new-type wrapper of a Validation having a valid outcome.
-pub struct Valid<'sm, 'v>(pub Validation<'sm, 'v>);
+// Nodes may be packed as FoundationDB tuples.
+pub mod tuple_pack;
 
-impl<'sm, 'v> Validation<'sm, 'v> {
-    /// Validate validates the given document against the given schema.
-    pub fn validate(
-        validator: &'v mut Validator<'sm>,
-        schema: &'v url::Url,
-        document: serde_json::Value,
-    ) -> Result<Self, json::schema::index::Error> {
-        validator.prepare(schema)?;
-        // Deserialization of Value cannot fail.
-        let span = json::de::walk(&document, validator).unwrap();
+pub mod inference;
 
-        Ok(Self {
-            schema,
-            document,
-            validator,
-            span,
-        })
+#[cfg(test)]
+mod test {
+
+    use super::{ArchivedNode, AsNode, HeapNode};
+    use serde_json::json;
+
+    #[test]
+    fn test_round_trip() {
+        let big_string = std::iter::repeat("a big string")
+            .take(30)
+            .collect::<Vec<_>>()
+            .join(" ");
+
+        let fixture = json!({
+            "numbers": [ 0x1111111111111111 as u64, -1234, 56.7891122334455],
+            "shared string": "shared string",
+            "some": {"bytes":"c29tZSBieXRlcw=="},
+            "null": null,
+            "nested": {
+                "true": true,
+                "false": false,
+                "two": 2,
+                "shared string": {"shared string": "shared string"},
+            },
+            "big string": big_string,
+            "small string": "smol", // Not de-duplicated because it's so small.
+        });
+
+        // We can deserialize into a Doc.
+        let alloc = HeapNode::new_allocator();
+        let dedup = HeapNode::new_deduper(&alloc);
+        let doc = HeapNode::from_serde(&fixture, &alloc, &dedup).unwrap();
+        insta::assert_debug_snapshot!(doc);
+
+        // The document can be archived with a stable byte layout.
+        let archive_buf = doc.to_archive();
+
+        let archive_buf_hexdump = hexdump::hexdump_iter(&archive_buf)
+            .map(|line| format!("{line}"))
+            .collect::<Vec<_>>()
+            .join("\n");
+        insta::assert_snapshot!(archive_buf_hexdump);
+
+        // We can directly serialize an ArchivedNode into a serde_json::Value,
+        // which exactly matches our original fixture.
+        let archived_doc = ArchivedNode::from_archive(&archive_buf);
+        let recovered = serde_json::to_value(archived_doc.as_node()).unwrap();
+        assert_eq!(fixture, recovered);
+
+        // The live document also serializes to an identical Value.
+        let recovered = serde_json::to_value(doc.as_node()).unwrap();
+        assert_eq!(fixture, recovered);
+
+        // A serde_json::Value can also be serialized as an AsNode.
+        let recovered = serde_json::to_value(fixture.as_node()).unwrap();
+        assert_eq!(fixture, recovered);
+
+        // Confirm number of bump-allocated bytes doesn't regress.
+        assert_eq!(alloc.allocated_bytes(), 1408);
     }
 
-    /// Ok returns returns FailedValidation if the validation failed, or Valid otherwise.
-    pub fn ok(self) -> Result<Valid<'sm, 'v>, FailedValidation> {
-        if !self.validator.invalid() {
-            return Ok(Valid(self));
-        }
+    #[test]
+    fn test_data_serialization() {
+        let bump = bumpalo::Bump::new();
 
-        let Self {
-            schema,
-            document,
-            validator,
-            span,
-        } = self;
+        let doc = HeapNode::Bytes(super::heap::BumpVec(
+            bumpalo::vec![in &bump; 8, 6, 7, 5, 3, 0, 9],
+        ));
+        let human_doc = serde_json::to_value(doc.as_node()).unwrap();
 
-        // Repeat the validation, but this time with FullContext for better error generation.
-        let mut full_validator =
-            json::validator::Validator::<Annotation, FullContext>::new(validator.schema_index());
-        full_validator.prepare(schema).unwrap();
-        let full_span = json::de::walk(&document, &mut full_validator).unwrap();
-
-        // Sanity check that we got the same validation result.
-        assert!(full_validator.invalid());
-        assert_eq!(span, full_span);
-
-        Err(FailedValidation {
-            document,
-            basic_output: json::validator::build_basic_output(full_validator.outcomes()),
-        })
+        insta::assert_debug_snapshot!(human_doc, @r###"String("CAYHBQMACQ==")"###);
     }
 }
-
-#[derive(Debug, serde::Serialize)]
-pub struct FailedValidation {
-    pub document: serde_json::Value,
-    pub basic_output: serde_json::Value,
-}
-
-impl std::fmt::Display for FailedValidation {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        serde_json::json!({
-            "document": &self.document,
-            "basic_output": &self.basic_output,
-        })
-        .fmt(f)
-    }
-}
-impl std::error::Error for FailedValidation {}
