+++ crates/doc/src/reduce/set.rs
@@ -1,83 +1,106 @@
-use super::{reduce_item, reduce_prop, Cursor, Error, Index, Location, Reducer, Result};
+use super::{
+    count_nodes, count_nodes_heap, reduce_item, reduce_prop, Cursor, Error, Index, Result,
+};
+use crate::{
+    dedup::Deduper,
+    heap::BumpVec,
+    lazy::{LazyArray, LazyDestructured, LazyField, LazyObject},
+    AsNode, Field, Fields, HeapField, HeapNode, LazyNode, Pointer,
+};
 use itertools::EitherOrBoth;
-use json::json_cmp_at;
-use serde_json::{Map, Value};
+use std::iter::Iterator;
 
-use super::strategy::Set;
+#[derive(serde::Serialize, serde::Deserialize, Debug)]
+#[serde(deny_unknown_fields, rename_all = "camelCase")]
+pub struct Set {
+    #[serde(default)]
+    pub key: Vec<Pointer>,
+}
 
 /// Permitted, destructured forms that set instances may take.
 /// Arrays are strictly ordered as "add", "intersect", "remove".
 /// This is sorted order: it's the order in which we'll reduce
 /// the RHS, and is the order in which we'll consume recursive
 /// annotation tokens of each property within the tape.
-#[derive(Debug)]
-pub enum Destructured {
+pub enum Destructured<'alloc, 'l, 'r, L: AsNode, R: AsNode> {
     Array {
-        props: [String; 3],
-        lhs: [Option<Vec<Value>>; 3],
-        rhs: [Option<Vec<Value>>; 3],
+        lhs: [Option<LazyArray<'alloc, 'l, L>>; 3],
+        rhs: [Option<LazyArray<'alloc, 'r, R>>; 3],
     },
     Object {
-        props: [String; 3],
-        lhs: [Option<Map<String, Value>>; 3],
-        rhs: [Option<Map<String, Value>>; 3],
+        lhs: [Option<LazyObject<'alloc, 'l, L>>; 3],
+        rhs: [Option<LazyObject<'alloc, 'r, R>>; 3],
     },
 }
 
-impl Destructured {
-    fn extract(loc: Location, lhs: Value, rhs: Value) -> Result<Self> {
+impl<'alloc, 'l, 'r, L: AsNode, R: AsNode> Destructured<'alloc, 'l, 'r, L, R> {
+    fn extract(
+        loc: json::Location,
+        lhs: LazyNode<'alloc, 'l, L>,
+        rhs: LazyNode<'alloc, 'r, R>,
+    ) -> Result<Self> {
         // Unwrap required Objects on each side.
-        let (lhs, rhs) = match (lhs, rhs) {
-            (Value::Object(lhs), Value::Object(rhs)) => (lhs, rhs),
-            (_, _) => return Err(Error::at(loc, Error::SetWrongType)),
+        let (lhs, rhs) = match (lhs.destructure(), rhs.destructure()) {
+            (LazyDestructured::Object(lhs), LazyDestructured::Object(rhs)) => (lhs, rhs),
+            (lhs, rhs) => return Err(Error::with_details(Error::SetWrongType, loc, lhs, rhs)),
         };
 
         // Extract "add", "intersect", and "remove" properties & values
         // from both sides, while dis-allowing both "intersect" and "remove"
         // on a single side. Inner object vs array types are disambiguated
         // for each property, and no other properties are permitted.
-        // |props| are property Strings obtained from either side, for re-use.
 
-        let mut props = [String::new(), String::new(), String::new()];
-        let mut rhs_arr = [None, None, None];
-        let mut rhs_obj = [None, None, None];
-        let mut lhs_arr = [None, None, None];
-        let mut lhs_obj = [None, None, None];
+        fn unpack<'alloc, 'n, N: AsNode>(
+            loc: json::Location,
+            obj: LazyObject<'alloc, 'n, N>,
+            arr_items: &mut [Option<LazyArray<'alloc, 'n, N>>; 3],
+            obj_items: &mut [Option<LazyObject<'alloc, 'n, N>>; 3],
+        ) -> Result<()> {
+            for field in obj.into_iter() {
+                let (property, value) = field.into_parts();
 
-        for (m, arr_items, obj_items) in &mut [
-            (lhs, &mut lhs_arr, &mut lhs_obj),
-            (rhs, &mut rhs_arr, &mut rhs_obj),
-        ] {
-            for (prop, item) in std::mem::take(m).into_iter() {
-                match (prop.as_ref(), item) {
-                    ("add", Value::Object(obj)) => {
-                        props[0] = prop;
+                // Collapse separate 'n and 'alloc lifetimes into one.
+                let property = match property {
+                    Ok(p) | Err(p) => p,
+                };
+
+                match (property, value.destructure()) {
+                    ("add", LazyDestructured::Object(obj)) => {
                         obj_items[0] = Some(obj);
                     }
-                    ("add", Value::Array(arr)) => {
-                        props[0] = prop;
+                    ("add", LazyDestructured::Array(arr)) => {
                         arr_items[0] = Some(arr);
                     }
-                    ("intersect", Value::Object(obj)) => {
-                        props[1] = prop;
+                    ("intersect", LazyDestructured::Object(obj)) => {
                         obj_items[1] = Some(obj);
                     }
-                    ("intersect", Value::Array(arr)) => {
-                        props[1] = prop;
+                    ("intersect", LazyDestructured::Array(arr)) => {
                         arr_items[1] = Some(arr);
                     }
-                    ("remove", Value::Object(obj)) if obj_items[1].is_none() => {
-                        props[2] = prop;
+                    ("remove", LazyDestructured::Object(obj)) if obj_items[1].is_none() => {
                         obj_items[2] = Some(obj);
                     }
-                    ("remove", Value::Array(arr)) if arr_items[1].is_none() => {
-                        props[2] = prop;
+                    ("remove", LazyDestructured::Array(arr)) if arr_items[1].is_none() => {
                         arr_items[2] = Some(arr);
                     }
-                    _ => return Err(Error::at(loc.push_prop(&prop), Error::SetWrongType)),
+                    (property, _) => {
+                        return Err(Error::with_location(
+                            Error::SetWrongType,
+                            loc.push_prop(property),
+                        ))
+                    }
                 }
             }
+            Ok(())
         }
+
+        let mut lhs_arr = [None, None, None];
+        let mut lhs_obj = [None, None, None];
+        unpack(loc, lhs, &mut lhs_arr, &mut lhs_obj)?;
+
+        let mut rhs_arr = [None, None, None];
+        let mut rhs_obj = [None, None, None];
+        unpack(loc, rhs, &mut rhs_arr, &mut rhs_obj)?;
 
         Ok(
             match (
@@ -85,15 +108,13 @@ impl Destructured {
                 lhs_obj.iter().any(Option::is_some) || rhs_obj.iter().any(Option::is_some),
             ) {
                 // Cannot mix array and object types.
-                (true, true) => return Err(Error::at(loc, Error::SetWrongType)),
+                (true, true) => return Err(Error::with_location(Error::SetWrongType, loc)),
 
                 (_, true) => Destructured::Object {
-                    props,
                     lhs: lhs_obj,
                     rhs: rhs_obj,
                 },
                 _ => Destructured::Array {
-                    props,
                     lhs: lhs_arr,
                     rhs: rhs_arr,
                 },
@@ -103,378 +124,575 @@ impl Destructured {
 }
 
 // Masks for defining merge outcomes and desired outcome filters.
+const NONE: u8 = 0;
 const LEFT: u8 = 1;
 const RIGHT: u8 = 2;
 const BOTH: u8 = 4;
 const UNION: u8 = 7;
 
 // Builder assists in building a set's constituent terms (add, intersect, remove).
-struct Builder<'i, 'l, 'a, 'k> {
-    tape: &'i mut Index<'a>,
-    loc: Location<'l>,
-    prune: bool,
-    key: &'k [String],
+struct Builder<'alloc, 'schema, 'tmp> {
+    tape: &'tmp mut Index<'schema>,
+    loc: json::Location<'tmp>,
+    full: bool,
+    key: &'schema [Pointer],
+    alloc: &'alloc bumpalo::Bump,
+    dedup: &'tmp Deduper<'alloc>,
 }
 
-impl Builder<'_, '_, '_, '_> {
+impl<'alloc> Builder<'alloc, '_, '_> {
     // Build the vector form of a term, as (LHS op1 SUB) op2 RHS.
     // If !naught, then op1 is LHS - SUB (eg, "remove all in SUB").
     // If naught, then op1 is LHS - SUB' (eg, "remove all *not* in SUB").
     //
     // The |mask| determines op2, which may be an intersection, union,
     // or set difference operation.
-    fn vec_term(
+    fn vec_term<L: AsNode, R: AsNode>(
         &mut self,
-        lhs: Option<Vec<Value>>,
-        sub: Option<&Vec<Value>>,
+        lhs: Option<LazyArray<'alloc, '_, L>>,
+        sub: Option<&LazyArray<'alloc, '_, R>>,
         naught: bool,
         mask: u8,
-        rhs: Option<Vec<Value>>,
-    ) -> Result<Option<Value>> {
-        // Flatten Option<Vec> into Vec.
-        let (lhs, rhs, tape_inc) = match (lhs, rhs) {
-            (Some(lhs), Some(rhs)) => (lhs, rhs, 1),
-            (Some(lhs), None) => (lhs, Vec::new(), 0),
-            (None, Some(rhs)) => (Vec::new(), rhs, 1),
-            (None, None) => return Ok(None),
+        rhs: Option<LazyArray<'alloc, '_, R>>,
+    ) -> Result<Option<HeapNode<'alloc>>> {
+        let Self {
+            tape,
+            loc,
+            full,
+            key,
+            alloc,
+            dedup,
+        } = self;
+
+        if rhs.is_some() {
+            **tape = &tape[1..]; // Consume |rhs| container.
+        } else if lhs.is_none() {
+            return Ok(None);
+        }
+
+        // Guess an output size and allocate its backing array.
+        let lhs_size = lhs.as_ref().map(LazyArray::len).unwrap_or_default();
+        let rhs_size = rhs.as_ref().map(LazyArray::len).unwrap_or_default();
+        let size_hint = match mask {
+            NONE => 0,
+            LEFT => lhs_size,
+            BOTH => lhs_size,
+            RIGHT => rhs_size,
+            UNION => lhs_size + rhs_size,
+            _ => unreachable!("invalid mask"),
         };
-        *self.tape = &self.tape[tape_inc..]; // Consume |rhs| container.
+        let mut arr = BumpVec::with_capacity_in(size_hint, alloc);
 
-        let empty = Vec::new();
-        let sub = sub.unwrap_or(&empty);
-
-        // Copy to allow multiple closures to reference |key| but not |self|.
-        let key = self.key;
-
-        let lhs_diff_sub =
-            itertools::merge_join_by(lhs.into_iter(), sub, |l, r| json_cmp_at(key, l, r))
-                .filter_map(|eob| match eob {
+        fn subtract<'i, 'alloc, 'l, 'r, L: AsNode, R: AsNode + 'r>(
+            key: &'i [Pointer],
+            left: impl Iterator<Item = LazyNode<'alloc, 'l, L>> + 'i,
+            right: impl Iterator<Item = &'r R> + 'i,
+            naught: bool,
+        ) -> Box<dyn Iterator<Item = LazyNode<'alloc, 'l, L>> + 'i> {
+            Box::new(
+                itertools::merge_join_by(left, right, |l, r| match l {
+                    LazyNode::Node(l) => Pointer::compare(key, *l, *r),
+                    LazyNode::Heap(l) => Pointer::compare(key, l, *r),
+                })
+                .filter_map(move |eob| match eob {
                     EitherOrBoth::Left(l) if !naught => Some(l),
                     EitherOrBoth::Both(l, _) if naught => Some(l),
                     _ => None,
-                });
+                }),
+            )
+        }
 
-        let v = itertools::merge_join_by(
+        let lhs = lhs.into_iter().flat_map(LazyArray::into_iter);
+        let lhs_diff_sub: Box<dyn Iterator<Item = LazyNode<_>>> = match sub {
+            Some(LazyArray::Node(arr)) => subtract(key, lhs, arr.iter(), naught),
+            Some(LazyArray::Heap(arr)) => subtract(key, lhs, arr.0.iter(), naught),
+            None => Box::new(lhs),
+        };
+
+        for eob in itertools::merge_join_by(
             lhs_diff_sub.enumerate(),
-            rhs.into_iter().enumerate(),
-            |(_, l), (_, r)| json_cmp_at(key, l, r),
-        )
-        .map(|eob| {
-            let outcome = match &eob {
-                EitherOrBoth::Left(_) => LEFT,
-                EitherOrBoth::Right(_) => RIGHT,
-                EitherOrBoth::Both(_, _) => BOTH,
+            rhs.into_iter().flat_map(LazyArray::into_iter).enumerate(),
+            |(_, l), (_, r)| LazyNode::compare(l, key, r),
+        ) {
+            match eob {
+                EitherOrBoth::Left((_, lhs)) if LEFT & mask != 0 => {
+                    arr.0.push(lhs.into_heap_node(alloc, dedup));
+                }
+                EitherOrBoth::Right((_, rhs)) if RIGHT & mask != 0 => {
+                    let rhs = rhs.into_heap_node(alloc, dedup);
+                    **tape = &tape[count_nodes_heap(&rhs)..];
+                    arr.0.push(rhs);
+                }
+                EitherOrBoth::Both(_, _) if BOTH & mask != 0 => {
+                    arr.0
+                        .push(reduce_item(*tape, *loc, *full, eob, alloc, dedup)?);
+                }
+                EitherOrBoth::Left(_) => {
+                    // Discard.
+                }
+                EitherOrBoth::Right((_, rhs)) | EitherOrBoth::Both(_, (_, rhs)) => {
+                    **tape = &tape[count_nodes(&rhs)..]; // Discard, but count nodes.
+                }
             };
-            Ok((outcome, reduce_item(self.tape, self.loc, self.prune, eob)?))
-        })
-        .filter_map(|r| match r {
-            Ok((outcome, value)) if outcome & mask != 0 => Some(Ok(value)),
-            Err(err) => Some(Err(err)),
-            Ok(_) => None,
-        })
-        .collect::<Result<Vec<Value>>>()?;
+        }
 
-        Ok(Some(Value::Array(v)))
+        Ok(Some(HeapNode::Array(arr)))
     }
 
     // Build the map form of a term. Behaves just like vec_term.
-    fn map_term(
+    fn map_term<L: AsNode, R: AsNode>(
         &mut self,
-        lhs: Option<Map<String, Value>>,
-        sub: Option<&Map<String, Value>>,
+        lhs: Option<LazyObject<'alloc, '_, L>>,
+        sub: Option<&LazyObject<'alloc, '_, R>>,
         naught: bool,
         mask: u8,
-        rhs: Option<Map<String, Value>>,
-    ) -> Result<Option<Value>> {
-        // Flatten Option<Map> into Map.
-        let (lhs, rhs, tape_inc) = match (lhs, rhs) {
-            (Some(lhs), Some(rhs)) => (lhs, rhs, 1),
-            (Some(lhs), None) => (lhs, Map::new(), 0),
-            (None, Some(rhs)) => (Map::new(), rhs, 1),
-            (None, None) => return Ok(None),
+        rhs: Option<LazyObject<'alloc, '_, R>>,
+    ) -> Result<Option<HeapNode<'alloc>>> {
+        let Self {
+            tape,
+            loc,
+            full,
+            key: _,
+            alloc,
+            dedup,
+        } = self;
+
+        if rhs.is_some() {
+            **tape = &tape[1..]; // Consume |rhs| container.
+        } else if lhs.is_none() {
+            return Ok(None);
+        }
+
+        // Guess an output size and allocate its backing array.
+        let lhs_size = lhs.as_ref().map(LazyObject::len).unwrap_or_default();
+        let rhs_size = rhs.as_ref().map(LazyObject::len).unwrap_or_default();
+        let size_hint = match mask {
+            NONE => 0,
+            LEFT => lhs_size,
+            BOTH => lhs_size,
+            RIGHT => rhs_size,
+            UNION => lhs_size + rhs_size,
+            _ => unreachable!("invalid mask"),
         };
-        *self.tape = &self.tape[tape_inc..]; // Consume |rhs| container.
+        let mut fields = BumpVec::with_capacity_in(size_hint, alloc);
 
-        let empty = Map::new();
-        let sub = sub.unwrap_or(&empty);
+        fn subtract<'i, 'alloc, 'l, 'r, L: AsNode, R: AsNode, F: Field<'r, R>>(
+            left: impl Iterator<Item = LazyField<'alloc, 'l, L>> + 'i,
+            right: impl Iterator<Item = F> + 'i,
+            naught: bool,
+        ) -> Box<dyn Iterator<Item = LazyField<'alloc, 'l, L>> + 'i> {
+            Box::new(
+                itertools::merge_join_by(left, right, |l, r| l.property().cmp(r.property()))
+                    .filter_map(move |eob| match eob {
+                        EitherOrBoth::Left(l) if !naught => Some(l),
+                        EitherOrBoth::Both(l, _) if naught => Some(l),
+                        _ => None,
+                    }),
+            )
+        }
 
-        let lhs_diff_sub =
-            itertools::merge_join_by(lhs.into_iter(), sub, |(l, _), (r, _)| l.cmp(r)).filter_map(
-                |eob| match eob {
-                    EitherOrBoth::Left(l) if !naught => Some(l),
-                    EitherOrBoth::Both(l, _) if naught => Some(l),
-                    _ => None,
-                },
-            );
+        let lhs = lhs.into_iter().flat_map(LazyObject::into_iter);
+        let lhs_diff_sub: Box<dyn Iterator<Item = LazyField<_>>> = match sub {
+            Some(LazyObject::Node(fields)) => subtract(lhs, fields.iter(), naught),
+            Some(LazyObject::Heap(fields)) => subtract(lhs, fields.0.iter(), naught),
+            None => Box::new(lhs),
+        };
 
-        let m = itertools::merge_join_by(lhs_diff_sub, rhs.into_iter(), |(l, _), (r, _)| l.cmp(r))
-            .map(|eob| {
-                let outcome = match &eob {
-                    EitherOrBoth::Left(_) => LEFT,
-                    EitherOrBoth::Right(_) => RIGHT,
-                    EitherOrBoth::Both(_, _) => BOTH,
-                };
-                Ok((outcome, reduce_prop(self.tape, self.loc, self.prune, eob)?))
-            })
-            .filter_map(|r| match r {
-                Ok((outcome, value)) if outcome & mask != 0 => Some(Ok(value)),
-                Err(err) => Some(Err(err)),
-                Ok(_) => None,
-            })
-            .collect::<Result<Map<_, _>>>()?;
+        for eob in itertools::merge_join_by(
+            lhs_diff_sub,
+            rhs.into_iter().flat_map(LazyObject::into_iter),
+            |l, r| l.property().cmp(r.property()),
+        ) {
+            match eob {
+                EitherOrBoth::Left(lhs) if LEFT & mask != 0 => {
+                    fields.0.push(lhs.into_heap_field(alloc, dedup));
+                }
+                EitherOrBoth::Right(rhs) if RIGHT & mask != 0 => {
+                    let rhs: HeapField = rhs.into_heap_field(alloc, dedup);
+                    **tape = &tape[count_nodes_heap(&rhs.value)..];
+                    fields.0.push(rhs);
+                }
+                EitherOrBoth::Both(_, _) if BOTH & mask != 0 => {
+                    fields
+                        .0
+                        .push(reduce_prop(*tape, *loc, *full, eob, alloc, dedup)?);
+                }
+                EitherOrBoth::Left(_) => {
+                    // Discard.
+                }
+                EitherOrBoth::Right(rhs) | EitherOrBoth::Both(_, rhs) => {
+                    let (_property, value) = rhs.into_parts();
+                    **tape = &tape[count_nodes(&value)..]; // Discard, but count nodes.
+                }
+            };
+        }
 
-        Ok(Some(Value::Object(m)))
+        Ok(Some(HeapNode::Object(fields)))
     }
 }
 
-impl Reducer for Set {
-    fn reduce(&self, cur: Cursor) -> Result<Value> {
-        let (tape, loc, prune, lhs, rhs) = match cur {
-            Cursor::Both {
-                tape,
-                loc,
-                prune,
-                lhs,
-                rhs,
-            } => (tape, loc, prune, lhs, rhs),
-            Cursor::Right {
-                tape,
-                loc,
-                prune,
-                rhs,
-            } => (tape, loc, prune, Value::Object(Map::new()), rhs),
-        };
+impl Set {
+    pub fn apply<'alloc, 'schema, L: AsNode, R: AsNode>(
+        &'schema self,
+        cur: Cursor<'alloc, 'schema, '_, '_, '_, L, R>,
+    ) -> Result<HeapNode<'alloc>> {
+        let Cursor {
+            tape,
+            loc,
+            lhs,
+            rhs,
+            alloc,
+            dedup,
+            full,
+        } = cur;
+
         *tape = &tape[1..]; // Consume object holding the set.
 
         let mut bld = Builder {
             tape,
             loc,
-            prune,
-            key: self.key.as_slice(),
+            full,
+            key: &self.key,
+            alloc,
+            dedup,
         };
-        let mut out = Map::with_capacity(2);
+        let mut out = BumpVec::with_capacity_in(2, alloc);
+
+        let add = bld.dedup.alloc_shared_string("add");
+        let intersect = bld.dedup.alloc_shared_string("intersect");
+        let remove = bld.dedup.alloc_shared_string("remove");
 
         match Destructured::extract(loc, lhs, rhs)? {
             // I,A reduce I,A
             Destructured::Array {
-                props: [add, intersect, _],
                 lhs: [la, Some(li), None],
                 rhs: [ra, Some(ri), None],
             } => {
                 // Reduce "add" as: (LA - RI') U RA.
-                if let Some(v) = bld.vec_term(la, Some(&ri), true, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.vec_term(la, Some(&ri), true, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: LI & RI.
-                if let (Some(v), false) = (
-                    bld.vec_term(Some(li), None, false, BOTH, Some(ri))?,
-                    bld.prune,
+                if let (Some(term), false) = (
+                    bld.vec_term(
+                        Some(li),
+                        None,
+                        false,
+                        if bld.full { NONE } else { BOTH },
+                        Some(ri),
+                    )?,
+                    bld.full,
                 ) {
-                    out.insert(intersect, v);
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // I,A reduce R,A
             Destructured::Array {
-                props: [add, intersect, _],
                 lhs: [la, Some(li), None],
                 rhs: [ra, None, rr],
             } => {
                 // Reduce "add" as: (LA - RR) U RA.
-                if let Some(v) = bld.vec_term(la, rr.as_ref(), false, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.vec_term(la, rr.as_ref(), false, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: LI - RR.
-                if let (Some(v), false) =
-                    (bld.vec_term(Some(li), None, false, LEFT, rr)?, bld.prune)
-                {
-                    out.insert(intersect, v);
+                if let (Some(term), false) = (
+                    bld.vec_term(
+                        Some(li),
+                        None,
+                        false,
+                        if bld.full { NONE } else { LEFT },
+                        rr,
+                    )?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // R,A reduce I,A
             Destructured::Array {
-                props: [add, intersect, _],
                 lhs: [la, None, lr],
                 rhs: [ra, Some(ri), None],
             } => {
                 // Reduce "add" as: (LA - RI') U RA.
-                if let Some(v) = bld.vec_term(la, Some(&ri), true, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.vec_term(la, Some(&ri), true, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: RI - LR.
-                if let (Some(v), false) =
-                    (bld.vec_term(lr, None, false, RIGHT, Some(ri))?, bld.prune)
-                {
-                    out.insert(intersect, v);
+                if let (Some(term), false) = (
+                    bld.vec_term(
+                        lr,
+                        None,
+                        false,
+                        if bld.full { NONE } else { RIGHT },
+                        Some(ri),
+                    )?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // R,A reduce R,A
             Destructured::Array {
-                props: [add, _, remove],
+                //props: [add, _, remove],
                 lhs: [la, None, lr],
                 rhs: [ra, None, rr],
             } => {
                 // Reduce "add" as: (LA - RR) U RA.
-                if let Some(v) = bld.vec_term(la, rr.as_ref(), false, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.vec_term(la, rr.as_ref(), false, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "remove" as: LR U RR.
-                if let (Some(v), false) = (bld.vec_term(lr, None, false, UNION, rr)?, bld.prune) {
-                    out.insert(remove, v);
+                if let (Some(term), false) = (
+                    bld.vec_term(lr, None, false, if bld.full { NONE } else { UNION }, rr)?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(remove) = term;
                 }
             }
 
             // I,A reduce I,A
             Destructured::Object {
-                props: [add, intersect, _],
                 lhs: [la, Some(li), None],
                 rhs: [ra, Some(ri), None],
             } => {
                 // Reduce "add" as: (LA - RI') U RA.
-                if let Some(v) = bld.map_term(la, Some(&ri), true, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.map_term(la, Some(&ri), true, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: LI & RI.
-                if let (Some(v), false) = (
-                    bld.map_term(Some(li), None, false, BOTH, Some(ri))?,
-                    bld.prune,
+                if let (Some(term), false) = (
+                    bld.map_term(
+                        Some(li),
+                        None,
+                        false,
+                        if bld.full { NONE } else { BOTH },
+                        Some(ri),
+                    )?,
+                    bld.full,
                 ) {
-                    out.insert(intersect, v);
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // I,A reduce R,A
             Destructured::Object {
-                props: [add, intersect, _],
                 lhs: [la, Some(li), None],
                 rhs: [ra, None, rr],
             } => {
                 // Reduce "add" as: (LA - RR) U RA.
-                if let Some(v) = bld.map_term(la, rr.as_ref(), false, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.map_term(la, rr.as_ref(), false, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: LI - RR.
-                if let (Some(v), false) =
-                    (bld.map_term(Some(li), None, false, LEFT, rr)?, bld.prune)
-                {
-                    out.insert(intersect, v);
+                if let (Some(term), false) = (
+                    bld.map_term(
+                        Some(li),
+                        None,
+                        false,
+                        if bld.full { NONE } else { LEFT },
+                        rr,
+                    )?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // R,A reduce I,A
             Destructured::Object {
-                props: [add, intersect, _],
+                //props: [add, intersect, _],
                 lhs: [la, None, lr],
                 rhs: [ra, Some(ri), None],
             } => {
                 // Reduce "add" as: (LA - RI') U RA.
-                if let Some(v) = bld.map_term(la, Some(&ri), true, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.map_term(la, Some(&ri), true, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "intersect" as: RI - LR.
-                if let (Some(v), false) =
-                    (bld.map_term(lr, None, false, RIGHT, Some(ri))?, bld.prune)
-                {
-                    out.insert(intersect, v);
+                if let (Some(term), false) = (
+                    bld.map_term(
+                        lr,
+                        None,
+                        false,
+                        if bld.full { NONE } else { RIGHT },
+                        Some(ri),
+                    )?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(intersect) = term;
                 }
             }
             // R,A reduce R,A
             Destructured::Object {
-                props: [add, _, remove],
+                //props: [add, _, remove],
                 lhs: [la, None, lr],
                 rhs: [ra, None, rr],
             } => {
                 // Reduce "add" as: (LA - RR) U RA.
-                if let Some(v) = bld.map_term(la, rr.as_ref(), false, UNION, ra)? {
-                    out.insert(add, v);
+                if let Some(term) = bld.map_term(la, rr.as_ref(), false, UNION, ra)? {
+                    *out.insert_mut(add) = term;
                 }
+
                 // Reduce "remove" as: LR U RR.
-                if let (Some(v), false) = (bld.map_term(lr, None, false, UNION, rr)?, bld.prune) {
-                    out.insert(remove, v);
+                if let (Some(term), false) = (
+                    bld.map_term(lr, None, false, if bld.full { NONE } else { UNION }, rr)?,
+                    bld.full,
+                ) {
+                    *out.insert_mut(remove) = term;
                 }
             }
 
-            _ => return Err(Error::at(loc, Error::SetWrongType)),
+            _ => return Err(Error::with_location(Error::SetWrongType, loc)),
         };
 
-        Ok(Value::Object(out))
+        Ok(HeapNode::Object(out))
     }
 }
 
 #[cfg(test)]
 mod test {
     use super::super::test::*;
-    use super::{Destructured, Location};
+    use super::Destructured;
+    use crate::LazyNode;
 
     #[test]
     fn test_destructure_cases() {
         use Destructured::{Array, Object};
-        let rt = Location::Root;
+        use LazyNode::Node;
+        let rt = json::Location::Root;
 
-        // Excercise add / intersect / remove on either side, with property collection.
-        let d = Destructured::extract(rt, json!({"add": []}), json!({"remove": []})).unwrap();
-        assert!(matches!(d, Array{
-                props: [add, _, remove],
+        // Exercise add / intersect / remove on either side, with property collection.
+        assert!(matches!(
+            Destructured::extract(rt, Node(&json!({"add": []})), Node(&json!({"remove": []})),)
+                .unwrap(),
+            Array {
                 lhs: [Some(_), None, None],
                 rhs: [None, None, Some(_)],
-            } if add == "add" && remove == "remove"));
+            }
+        ));
 
-        let d = Destructured::extract(rt, json!({"remove": []}), json!({"intersect": []})).unwrap();
-        assert!(matches!(d, Array{
-                props: [_, intersect, remove],
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({"remove": []})),
+                Node(&json!({"intersect": []}))
+            )
+            .unwrap(),
+            Array {
                 lhs: [None, None, Some(_)],
                 rhs: [None, Some(_), None],
-            } if intersect == "intersect" && remove == "remove"));
+            }
+        ));
 
-        let d = Destructured::extract(rt, json!({"intersect": []}), json!({"add": []})).unwrap();
-        assert!(matches!(d, Array{
-                props: [add, intersect, _],
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({"intersect": []})),
+                Node(&json!({"add": []}))
+            )
+            .unwrap(),
+            Array {
                 lhs: [None, Some(_), None],
                 rhs: [Some(_), None, None],
-            } if add == "add" && intersect == "intersect"));
+            }
+        ));
 
-        let d = Destructured::extract(rt, json!({"add": {}}), json!({"remove": {}})).unwrap();
-        assert!(matches!(d, Object{
-                props: [add, _, remove],
+        assert!(matches!(
+            Destructured::extract(rt, Node(&json!({"add": {}})), Node(&json!({"remove": {}})))
+                .unwrap(),
+            Object {
                 lhs: [Some(_), None, None],
                 rhs: [None, None, Some(_)],
-            } if add == "add" && remove == "remove"));
+            }
+        ));
 
-        let d = Destructured::extract(rt, json!({"remove": {}}), json!({"intersect": {}})).unwrap();
-        assert!(matches!(d, Object{
-                props: [_, intersect, remove],
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({"remove": {}})),
+                Node(&json!({"intersect": {}}))
+            )
+            .unwrap(),
+            Object {
+                //props: [_, intersect, remove],
                 lhs: [None, None, Some(_)],
                 rhs: [None, Some(_), None],
-            } if intersect == "intersect" && remove == "remove"));
+            }
+        ));
 
-        let d = Destructured::extract(rt, json!({"intersect": {}}), json!({"add": {}})).unwrap();
-        assert!(matches!(d, Object{
-                props: [add, intersect, _],
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({"intersect": {}})),
+                Node(&json!({"add": {}}))
+            )
+            .unwrap(),
+            Object {
+                //props: [add, intersect, _],
                 lhs: [None, Some(_), None],
                 rhs: [Some(_), None, None],
-            } if add == "add" && intersect == "intersect"));
+            }
+        ));
 
         // Either side may be empty.
-        let d = Destructured::extract(rt, json!({}), json!({"add": {}, "remove": {}})).unwrap();
-        assert!(matches!(d, Object{
-            lhs: [None, None, None],
-            rhs: [Some(_), None, Some(_)],
-            ..
-        }));
-        let d = Destructured::extract(rt, json!({"add": [], "remove": []}), json!({})).unwrap();
-        assert!(matches!(d, Array{
-            lhs: [Some(_), None, Some(_)],
-            rhs: [None, None, None],
-            ..
-        }));
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({})),
+                Node(&json!({"add": {}, "remove": {}}))
+            )
+            .unwrap(),
+            Object {
+                lhs: [None, None, None],
+                rhs: [Some(_), None, Some(_)],
+            }
+        ));
+        assert!(matches!(
+            Destructured::extract(
+                rt,
+                Node(&json!({"add": [], "remove": []})),
+                Node(&json!({}))
+            )
+            .unwrap(),
+            Array {
+                lhs: [Some(_), None, Some(_)],
+                rhs: [None, None, None],
+            }
+        ));
 
         // Cases that fail:
 
         // Mixed types within a side.
-        Destructured::extract(rt, json!({"add": {}, "intersect": []}), json!({})).unwrap_err();
+        assert!(Destructured::extract(
+            rt,
+            Node(&json!({"add": {}, "intersect": []})),
+            Node(&json!({})),
+        )
+        .is_err());
         // Mixed types across sides.
-        Destructured::extract(rt, json!({"add": {}}), json!({"intersect": []})).unwrap_err();
+        assert!(Destructured::extract(
+            rt,
+            Node(&json!({"add": {}})),
+            Node(&json!({"intersect": []}))
+        )
+        .is_err());
         // Both "intersect" and "remove" on a side.
-        Destructured::extract(rt, json!({"intersect": [], "remove": []}), json!({})).unwrap_err();
+        assert!(Destructured::extract(
+            rt,
+            Node(&json!({"intersect": [], "remove": []})),
+            Node(&json!({}))
+        )
+        .is_err());
         // Not an object.
-        Destructured::extract(rt, json!({"intersect": []}), json!(42)).unwrap_err();
+        assert!(
+            Destructured::extract(rt, Node(&json!({"intersect": []})), Node(&json!(42))).is_err()
+        );
     }
 
     #[test]
@@ -549,6 +767,10 @@ mod test {
                     rhs: json!({"add": [[33, 1]]}),
                     expect: Ok(json!({"add": [[22, 2], [33, 2]]})),
                 },
+                Partial {
+                    rhs: json!({"remove": [[33]]}),
+                    expect: Ok(json!({"add": [[22, 2]], "remove": [[33]]})),
+                },
             ],
         )
     }
@@ -615,6 +837,10 @@ mod test {
                 Full {
                     rhs: json!({"add": {"33": 1}}),
                     expect: Ok(json!({"add": {"22": 2, "33": 2}})),
+                },
+                Partial {
+                    rhs: json!({"remove": {"33":0}}),
+                    expect: Ok(json!({"add": {"22":2}, "remove": {"33":0}})),
                 },
             ],
         )
