+++ crates/doc/src/ser.rs
