+++ crates/doc/tests/reduce_annotations_test.rs
@@ -4,7 +4,9 @@ extern crate quickcheck;
 #[macro_use(quickcheck)]
 extern crate quickcheck_macros;
 
-use doc::{reduce, Schema, SchemaIndexBuilder, Validation, Validator};
+use doc::{
+    compare, reduce, AsNode, HeapNode, LazyNode, Schema, SchemaIndexBuilder, Validation, Validator,
+};
 use itertools::{EitherOrBoth, Itertools};
 use json::schema::build::build_schema;
 use serde::Deserialize;
@@ -62,6 +64,8 @@ fn test_validate_then_reduce() {
     let index = index.into_index();
 
     let mut validator = Validator::new(&index);
+    let alloc = HeapNode::new_allocator();
+    let dedup = HeapNode::new_deduper(&alloc);
 
     let cases = vec![
         (json!({"lww": "one"}), json!({"lww": "one"})),
@@ -135,15 +139,33 @@ fn test_validate_then_reduce() {
         ),
     ];
 
-    let mut lhs: Option<Value> = None;
+    let mut lhs: Option<HeapNode<'_>> = None;
+
     for (rhs, expect) in cases {
-        let rhs = Validation::validate(&mut validator, &curi, rhs)
+        let rhs_valid = Validation::validate(&mut validator, &curi, &rhs)
             .unwrap()
             .ok()
             .unwrap();
-        let reduced = reduce::reduce(lhs, rhs, true).unwrap();
-        assert_eq!(&reduced, &expect);
-        lhs = Some(reduced);
+
+        let reduced = match lhs {
+            Some(lhs) => reduce::reduce(
+                LazyNode::Heap(lhs),
+                LazyNode::Node(&rhs),
+                rhs_valid,
+                &alloc,
+                &dedup,
+                true,
+            )
+            .unwrap(),
+            None => HeapNode::from_node(rhs.as_node(), &alloc, &dedup),
+        };
+
+        assert_eq!(
+            compare(&reduced, &expect),
+            std::cmp::Ordering::Equal,
+            "reduced: {reduced:?} expected: {expect:?}",
+        );
+        lhs = Some(reduced)
     }
 }
 
@@ -242,6 +264,7 @@ fn test_qc_set_array(mut seq: Vec<(bool, Vec<u8>, Vec<u8>)>) -> bool {
 #[derive(Deserialize, Debug)]
 #[serde(deny_unknown_fields)]
 struct TestArray {
+    #[serde(default)]
     add: Vec<(u8, u32)>,
 }
 
@@ -343,10 +366,16 @@ fn test_qc_set_map(seq: Vec<(bool, Vec<u8>, Vec<u8>)>) -> bool {
 #[derive(Deserialize, Debug)]
 #[serde(deny_unknown_fields)]
 struct TestMap {
+    #[serde(default)]
     add: BTreeMap<String, u32>,
 }
 
-fn reduce_tree(validator: &mut Validator, curi: &Url, mut docs: Vec<Value>) -> Value {
+fn reduce_tree(validator: &mut Validator, curi: &Url, docs: Vec<Value>) -> Value {
+    let alloc = HeapNode::new_allocator();
+    let dedup = HeapNode::new_deduper(&alloc);
+
+    let mut docs = docs.iter().map(LazyNode::Node).collect::<Vec<_>>();
+
     // Iteratively reduce |docs| by walking it in chunked windows, producing
     // a new Value for each chunk. Intuitively, we're reducing |docs| by
     // interpreting it as a tree and ascending from leaf to root (as opposed
@@ -360,19 +389,28 @@ fn reduce_tree(validator: &mut Validator, curi: &Url, mut docs: Vec<Value>) -> V
             .into_iter()
             .enumerate()
             .map(|(n, chunk)| {
-                let mut lhs: Option<Value> = None;
+                let mut lhs: Option<LazyNode<Value>> = None;
 
                 for rhs in chunk {
-                    let rhs = Validation::validate(validator, curi, rhs)
-                        .unwrap()
-                        .ok()
-                        .unwrap();
-                    lhs = Some(reduce::reduce(lhs, rhs, n == 0).unwrap());
+                    let rhs_valid = rhs.validate_ok(validator, curi).unwrap().unwrap();
+
+                    lhs = Some(match lhs {
+                        Some(lhs) => LazyNode::Heap(
+                            reduce::reduce(lhs, rhs, rhs_valid, &alloc, &dedup, n == 0).unwrap(),
+                        ),
+                        None => rhs,
+                    });
                 }
                 lhs.unwrap()
             })
             .collect();
     }
 
-    docs.into_iter().next().unwrap()
+    let root = docs
+        .into_iter()
+        .next()
+        .unwrap()
+        .into_heap_node(&alloc, &dedup);
+
+    serde_json::to_value(&root.as_node()).unwrap()
 }
