+++ crates/flowctl-admin/Cargo.toml
@@ -29,6 +29,7 @@ itertools = { workspace = true }
 reqwest = { workspace = true }
 serde_json = { workspace = true }
 superslice = { workspace = true }
+tempfile = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
